backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.68",
    "celery>=5.6.2",
    "fastapi>=0.135.1",
    "fastmcp>=3.1.0",
    "langchain>=1.2.12",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.11",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.2",
    "pinecone>=8.1.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
